+++ src/string_buffer.c
@@ -377,7 +377,7 @@ int buffer_printf(string_buffer_t *buffer, size_t buffer_row, char *buf, size_t 
     }
 
     int set_old_value = 0;
-    int  written = 0;
+    size_t  written = 0;
     int tmp = 0;
     ptrdiff_t str_it_width = 0;
     const CHAR_TYPE *beg = NULL;
@@ -405,7 +405,7 @@ int buffer_printf(string_buffer_t *buffer, size_t buffer_row, char *buf, size_t 
     set_old_value = 0;
     CHCK_RSLT_ADD_TO_WRITTEN(SNPRINT_N_STRINGS(buf + written,  total_buf_len - written, (content_width - (size_t)str_it_width), SPACE_CHAR));
     CHCK_RSLT_ADD_TO_WRITTEN(SNPRINT_N_STRINGS(buf + written, total_buf_len - written, right, SPACE_CHAR));
-    return written;
+    return (int)written;
 
 clear:
     if (set_old_value)
@@ -474,7 +474,7 @@ int buffer_wprintf(string_buffer_t *buffer, size_t buffer_row, wchar_t *buf, siz
     }
 
     int set_old_value = 0;
-    int  written = 0;
+    size_t  written = 0;
     int tmp = 0;
     ptrdiff_t str_it_width = 0;
     const CHAR_TYPE *beg = NULL;
@@ -502,7 +502,7 @@ int buffer_wprintf(string_buffer_t *buffer, size_t buffer_row, wchar_t *buf, siz
     set_old_value = 0;
     CHCK_RSLT_ADD_TO_WRITTEN(SNPRINT_N_STRINGS(buf + written,  total_buf_len - written, (content_width - (size_t)str_it_width), SPACE_CHAR));
     CHCK_RSLT_ADD_TO_WRITTEN(SNPRINT_N_STRINGS(buf + written, total_buf_len - written, right, SPACE_CHAR));
-    return written;
+    return (int)written;
 
 clear:
     if (set_old_value)
